[finance_report_vision] Split reporting.py (2143 lines) into a reporting/ package
Behaviour-preserving modularization of the single biggest domain file. The flat
2143-line module became a package of per-statement submodules, each well under
1000 lines; the public import surface is unchanged.

- reporting/_core.py: shared aggregation core (balances SQL, fx map, account
  lines, net-income SQL).
- reporting/{balance_sheet,income_statement,cash_flow,net_worth,lineage}.py:
  one statement/view per file.
- reporting/{manual_valuation,portfolio_market,internal_transfer}.py: the
  balance-sheet adjustment plugins (manual_valuation is the legacy net-worth
  side-channel's home, now isolated).
- reporting/__init__.py: re-exports the public functions + the reporting_calc
  primitives (AC25.1.1 single-owner contract preserved).

No production behaviour change. Tests that monkeypatched fx primitives /
internal aggregators at the flat module path are repointed to the submodule
where each name is looked up. 257 reporting tests pass; full collection clean.
